refactor(architecture): refactor chats
