+++ router/chats/handlers/handleDeleteChat.js
@@ -1,0 +1,28 @@
+const models = require('../../../models');
+const Chat = models.Chat;
+
+const handleDeleteChat = async (req, res) => {
+  const { id } = req.params;
+
+  try {
+    const chat = await Chat.findOne({
+      where: {
+        id,
+      },
+      include: [
+        {
+          model: User,
+        },
+      ],
+    });
+
+    const notifyUsers = chat.Users.map((user) => user.id);
+
+    await chat.destroy();
+    return res.json({ chatId: id, notifyUsers });
+  } catch (e) {
+    return res.status(500).json({ status: 'Error', message: e.message });
+  }
+};
+
+module.exports = handleDeleteChat;
