+++ router/chats/swagger/allChats.swagger.js
@@ -1,0 +1,12 @@
+/**
+ * @swagger
+ * /chats:
+ *   get:
+ *     summary: Get all chats for current user
+ *     tags: [Chats]
+ *     security:
+ *       - bearerAuth: []
+ *     responses:
+ *       200:
+ *         description: List of chats
+ */
