+++ router/chats/swagger/chatMessages.swagger.js
@@ -1,0 +1,12 @@
+/**
+ * @swagger
+ * /chats/messages:
+ *   get:
+ *     summary: Get recent messages for the current user's chats
+ *     tags: [Chats]
+ *     security:
+ *       - bearerAuth: []
+ *     responses:
+ *       200:
+ *         description: List of recent messages
+ */
